pyproject.toml
[project]
name = "metagene"
version = "0.0.0.dev54"
description = "Metagene Profiling Analysis and Visualization"
authors = [{ name = "Ye Chang", email = "yech1990@gmail.com" }]
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
keywords = [
    "DNA",
    "RNA",
    "metagene",
    "biology",
]
dependencies = [
    "matplotlib>=3.0.0,<4",
    "pyranges1>=1.0.13",
    "rich-click>=1.6.1,<2",
    "asciichartpy>=1.5.25,<2",
    "polars>=1.30.0",
    "pyarrow",
    "pandas>=2.2.3",
]

[project.urls]
Repository = "https://github.com/y9c/metagene"

[project.scripts]
metagene = "metagene.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["metagene"]

[tool.black]
line-length = 79
